Move CustomPrintStream to main code
