+++ Lushu/src/test/kotlin/lushu/Lushu/LogGenerator/LogGeneratorTest.kt
@@ -1,5 +1,6 @@
 package lushu.Lushu.LogGenerator
 
+import lushu.Interceptor.LushuPrintStream
 import org.junit.jupiter.api.Test
 import java.io.PrintStream
 import java.nio.file.Paths
@@ -7,7 +8,7 @@ import java.nio.file.Paths
 class LogGeneratorTest {
     @Test
     fun run() {
-        System.setOut(CustomPrintStream(System.out))
+        System.setOut(LushuPrintStream(System.out))
         val lg = LogGenerator(
             Paths.get("src", "test", "fixtures", "logs")
                 .toAbsolutePath().toString()
